mudei entrypont.sh para o mesmo conteudo do codigo que fiz o fork

+++ standalone/entrypoint.sh
@@ -1,16 +1,27 @@
 #!/bin/bash
 # Install extra deps from /opt/extra-deps.txt if it exists
-pip install -r /opt/extra-dependencies.txt
+if [ -f /opt/extra-dependencies.txt ]; then
+    pip install -r /opt/extra-dependencies.txt
+fi
+
 cd /opt/django-helpdesk/standalone/
+if python manage.py showmigrations | grep '\[ \]\|^[a-z]' | grep '[  ]' -B 1; then
+    python manage.py migrate --noinput                # Apply database migrations
+fi
 
 # Starting cron to check emails
 printenv > /etc/env
 env | awk -F= '{printf "export %s=\"%s\"\n", $1, $2}' > /etc/env
 cron &&
-python manage.py migrate --noinput                # Apply database migrations
-python manage.py createsuperuser --noinput
-python manage.py collectstatic --noinput
-
 # Start Gunicorn processes
 echo Starting Gunicorn.
-gunicorn config.wsgi:application --bind 0.0.0.0:8000
+exec gunicorn standalone.config.wsgi:application \
+	   --name django-helpdesk \
+	   --bind 0.0.0.0:${GUNICORN_PORT:-"8000"} \
+	   --workers ${GUNICORN_NUM_WORKERS:-"6"} \
+	   --timeout ${GUNICORN_TIMEOUT:-"60"} \
+	   --preload \
+	   --log-level=debug \
+	   --log-file=- \
+	   --access-logfile=- \
+	   "$@"
